Introduce SidedEnergyContainer with abstract get and set
